See Issue #6763