@JohannesGaessler 2024-04-22
Some single GPU test results:

| GPU         | Model           | Test  | t/s master | t/s PR | Speedup |
|-------------|-----------------|-------|------------|--------|---------|
| 1x RTX 4090 | LLaMA 3 8b f16  | pp512 |       9776 |   9777 |    1.00 |
| 1x RTX 4090 | LLaMA 3 8b f16  | tg128 |      54.92 |  55.96 |    1.02 |
| 1x RTX 4090 | LLaMA 3 8b q4_0 | pp512 |       7387 |   7385 |    1.00 |
| 1x RTX 4090 | LLaMA 3 8b q4_0 | tg128 |     137.50 | 146.02 |    1.06 |
| 1x P40      | LLaMA 2 7b q4_0 | pp512 |        908 |    909 |    1.00 |
| 1x P40      | LLaMA 2 7b q4_0 | tg128 |      53.84 |  51.67 |    0.96 |
| 1x RX 6800  | LLaMA 2 7b q4_0 | pp512 |        712 |    713 |    1.00 |
| 1x RX 6800  | LLaMA 2 7b q4_0 | tg128 |      60.25 |  60.66 |    1.01 |

There is a performance regression for P40. I was not able to run multi GPU tests because `llama-bench` crashes. Error with 6x RTX 4090:

```
CUDA error: operation not supported
  current device: 2, in function ggml_backend_cuda_graph_compute at ggml-cuda.cu:2558
  cudaGraphInstantiate(&cudaGraph.instance, cudaGraph.graph, __null, __null, 0)
GGML_ASSERT: ggml-cuda.cu:60: !"CUDA error"
```

Error with 3x P40:

```
CUDA error: an illegal memory access was encountered
  current device: 2, in function ggml_backend_cuda_synchronize at ggml-cuda.cu:2403
  cudaStreamSynchronize(cuda_ctx->stream())
GGML_ASSERT: ggml-cuda.cu:60: !"CUDA error"
```