@jdecourval 2024-04-23
> > With your changes, it now works with ROCm HIP (with patch below), but it is slower, making it likely not worth enabling it on that platform. I'm using a RX 6700 XT.
> > model	size	params	backend	ngl	test	t/s master	t/s PR	speedup
> > llama 7B Q4_K - Medium	3.80 GiB	6.74 B	ROCm	99	pp 512	850.47 ± 0.18	850.52 ± 0.08	1.00
> > llama 7B Q4_K - Medium	3.80 GiB	6.74 B	ROCm	99	tg 128	54.05 ± 0.02	52.86 ± 0.05	0.98
> > The patch can be found here: https://gist.github.com/ardfork/a223a10d20961707e7b5f3ee0b76c7d5, didn't want to bloat your PR comments with a wall of text that will be useless for most people reading it.
> 
> It work decently on RDNA3 through. I forget to record the value but it's about 2% faster for token generation.

Here's on my 7900xtx

| model                          |       size |     params | backend    | ngl | test       |              t/s before | t/s after |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: | ------: |
| command-r 35B IQ3_XS - 3.3 bpw |  15.65 GiB |    37.08 B | ROCm       |  99 | pp 512     |    737.72 ± 1.57 | 738.31 ± 1.48 |
| command-r 35B IQ3_XS - 3.3 bpw |  15.65 GiB |    37.08 B | ROCm       |  99 | tg 128     |     29.00 ± 0.05 | 28.94 ± 0.02 |
| llama 8B Q5_K - Medium         |   4.78 GiB |     7.24 B | ROCm       |  99 | pp 512     |   3219.19 ± 8.14 | 3218.28 ± 8.52 |
| llama 8B Q5_K - Medium         |   4.78 GiB |     7.24 B | ROCm       |  99 | tg 128     |     87.13 ± 0.06 | 87.37 ± 0.06 |
| llama 8x7B IQ3_XXS - 3.0625 bpw |  33.27 GiB |    91.80 B | ROCm       |  99 | pp 512     |   1165.00 ± 4.31 | 1165.52 ± 4.11|
| llama 8x7B IQ3_XXS - 3.0625 bpw |  33.27 GiB |    91.80 B | ROCm       |  99 | tg 128     |     55.08 ± 0.04 | 55.39 ± 0.04 |
| llama 8B Q6_K                  |   5.53 GiB |     7.24 B | ROCm       |  99 | pp 512     |  3273.69 ± 10.66 | 3268.28 ± 10.24 |
| llama 8B Q6_K                  |   5.53 GiB |     7.24 B | ROCm       |  99 | tg 128     |     87.24 ± 0.07 | 87.51 ± 0.05 |

Over those 4 models, your PR yields a significant, but very small average TG speedup of 0.2%.