@slaren 2024-05-08
Fixed all the issues I could find, I think this is ready to be merged, but more testing would be welcome.

I get this performance on my system (with flash attention enabled):

| GPU         | Model          |   Model Size [GiB] |  Test   |   t/s master |   t/s ag/ag_cuda_graphs |   Speedup |
|:------------|:---------------|-------------------:| :-------|-------------:|------------------------:|----------:|
| RTX 3090 Ti | llama 7B F16   |              12.55 |  tg128  |        57.05 |                   59.80 |      1.05 |
| RTX 3090 Ti | llama 7B Q4_0  |               3.56 |  tg128  |       138.58 |                  163.95 |      1.18 |
| RTX 3090 Ti | llama 13B Q4_0 |               6.86 |  tg128  |        85.43 |                   95.73 |      1.12 |
| RTX 3090 Ti | llama 30B Q4_0 |              17.09 |  tg128  |        39.32 |                   42.72 |      1.09 |

@agray3 I was wondering if it would be possible to use `cudaStreamBeginCaptureToGraph` to keep track of which nodes in the ggml graph correspond to which nodes in the cuda graph. In other words, is it possible to query the nodes of the cuda graph as they are captured, or only after `cudaStreamEndCapture`? If so, I think this could allow us to implement a more reliable update mechanism in the future.